refactor(speech-core): split TTS runtime
